Once again fixing some string fontification
It looks like the problem was the use of `end` instead of
`(treesit-node-end node)` to distinguish the boundary of what should
be fontified.

As part of this, I'm also cleaning up how the fontification works: we
are no longer using an override to fontify the `$`, `${`, and `}`
symbols, but rather letting the normal fontification do its thing.