When running Gradle command on the top-level project, don't run commands in subprojects
The specific use case for this is when running something like
`./gradlew test --tests=SomeTest`. Before this change, this would try
to run `SomeTest` in every subproject, which would fail, because it
probably didn't exist.

With this change, we now run `:test` instead of `test`, which causes
it to only run in the top-level project.